requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0

# LangChain and LLM
langchain==0.1.20
langchain-google-genai==1.0.8
langchain-experimental==0.0.64
langchain-community==0.0.38

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.21

# Vector operations
numpy==1.24.3
faiss-cpu==1.7.4

# Configuration management
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Utilities
requests==2.31.0
typing-extensions==4.8.0